% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basicRelationships}
\alias{basicRelationships}
\title{Basic relationships}
\format{
A data frame with 12 rows and 7 columns. The last column (\code{pos}) is
used internally for placing labels on triangle plots.
}
\usage{
basicRelationships
}
\description{
A data frame containing kinship and kappa coefficients for some commonly used
relationships.
}
\keyword{datasets}
