% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacquard.R
\name{jacquard}
\alias{jacquard}
\alias{detailed2condensed}
\title{Jacquard coefficients}
\usage{
jacquard(x, ids, detailed = FALSE, named = FALSE, verbose = FALSE)

detailed2condensed(d, named = FALSE)
}
\arguments{
\item{x}{A pedigree in the form of a \code{\link[pedtools:ped]{pedtools::ped}} object.}

\item{ids}{A vector of two ID labels.}

\item{detailed}{A logical. If FALSE (default), the 9 condensed coefficients
are computed; otherwise the 15 detailed identity coefficients.}

\item{named}{A logical, indicating if the output vector should be named.}

\item{verbose}{A logical.}

\item{d}{A numerical vector of length 15.}
}
\value{
A numeric vector of length 9 or 15 (depending on \code{detailed}). The
coefficients are given in the order used by Jacquard (1974).
}
\description{
This function calculates the pairwise identity coefficients defined by
Jacquard (1974). Unlike \code{\link[=condensedIdentity]{condensedIdentity()}} (and
\code{identity::identity.coefs()}), this function also computes the 15 \emph{detailed}
identity coefficients. If \code{detailed = FALSE} the output is equivalent to
\code{\link[=condensedIdentity]{condensedIdentity()}}, which is usually faster.
}
\details{
Karigl (1981) gave the first recursive algorithm for the 9 condensed Jacqaurd
coefficients, based on certain \emph{generalised kinship coefficients}. Karigl's
method is implemented in \code{\link[=condensedIdentity]{condensedIdentity()}}.

Weeks & Lange (1988) suggested a broader and more natural generalisation of
kinship coefficients, and a recursive algorithm for computing them. This is
implemented in \code{\link[=generalisedKinship]{generalisedKinship()}}.

Lange & Sinsheimer (1992) described an even further generalisation of kinship
coefficients, allowing a mix of deterministic and random sampling of alleles.
This is also implemented in \code{\link[=generalisedKinship]{generalisedKinship()}}.

In the same paper, Lange & Sinsheimer (1992) used the new generalisations to
give (i) an alternative algorithm for the 9 condensed identity coefficients,
and (ii) an algorithm for the 15 detailed coefficients. Both of these are
implemented in \code{jacquard()}.

Both the condensed and detailed coefficients are given in the orders used by
Jacquard (1974). The function \code{detailed2condensed} converts from detailed
coefficients (d1, ... d15) to condensed ones (D1, ..., D9) using the
following relations:
\itemize{
\item D1 = d1
\item D2 = d6
\item D3 = d2 + d3
\item D4 = d7
\item D5 = d4 + d5
\item D6 = d8
\item D7 = d9 + d12
\item D8 = d10 + d11 + d13 + d14
\item D9 = d15
}
}
\examples{
x = fullSibMating(1)
ids = 5:6

# Condensed coefficient
j = jacquard(x, ids)
stopifnot(identical(j, condensedIdentity(x, ids)))

# Detailed coefficients
jdet = jacquard(x, ids, detailed = TRUE)
jdet
stopifnot(identical(j, detailed2condensed(jdet)))

}
\references{
\itemize{
\item Jacquard, A. (1974). The Genetic Structure of Populations. Springer.
\item Karigl, G. (1981). A recursive algorithm for the calculation of identity
coefficients. Ann. Hum. Genet.
\item Weeks, D.E. & Lange, K. (1988). The affected-pedigree-member method of
linkage analysis. Am. J. Hum. Genet
\item Lange, K. & Sinsheimer, J.s. (1992). Calculation of genetic identity
coefficients. Ann. Hum. Genet.
}
}
\seealso{
\code{\link[=condensedIdentity]{condensedIdentity()}}, \code{\link[=idcoefs]{idcoefs()}}
}
