% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jicaque}
\alias{jicaque}
\title{Jicaque pedigree}
\format{A data frame with 22 rows and four columns:
\itemize{
\item \code{id}  : individual ID
\item \code{fid} : father's ID (or 0 if not included)
\item \code{mid} : mother's ID (or 0 if not included)
\item \code{sex} : Gender codes, where 1 = male and 2 = female
}}
\usage{
jicaque
}
\description{
A data frame describing a pedigree from the Jicaque tribe, studied by Chapman
and Jacquard (1971).
}
\references{
Chapman, A.M and Jacquard, A. (1971). Un isolat d'Amerique
Centrale: les Indiens Jicaques de Honduras. In Genetique et Population.
Paris: Presses Universitaires de France.
}
\keyword{datasets}
