% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappa}
\alias{kappa}
\title{IBD (kappa) coefficients}
\usage{
kappa(x, ids, ...)
}
\arguments{
\item{x}{A pedigree in the form of a \code{\link[pedtools:ped]{pedtools::ped}} object}

\item{ids}{A character (or coercible to character) of length 2, containing ID
labels of two pedigree members}

\item{...}{Arguments passed on to \code{condensedIdentity}}
}
\value{
The probability vector \eqn{(\kappa_0, \kappa_1, \kappa_2)}{(\kappa0,
  \kappa1, \kappa2)}. If any of the two individuals are inbred, \code{c(NA, NA, NA)} is returned.
}
\description{
Computes the three IBD coefficients summarising the relationship between two
non-inbred individuals.
}
\details{
For any pair of non-inbred individuals A and B, their genetic relationship
can be summarized by the IBD coefficients \eqn{(\kappa_0, \kappa_1,
\kappa_2)}{(\kappa0, \kappa1, \kappa2)}, where \deqn{\kappa_i = P(A and B
share i alleles IBD at a random autosomal locus)}

The current implementation calls \code{\link[=condensedIdentity]{condensedIdentity()}} and returns the three last
coefficients in the reverse order.

The function checks if any of the \code{ids} individuals are inbred. If so, a
message is printed to the screen, and \code{c(NA, NA, NA)} is returned.
}
\examples{
# Siblings
x = nuclearPed(2)
k = kappa(x, 3:4)
stopifnot(identical(k, c(.25, .5, .25)))

# Quad half first cousins
x = quadHalfFirstCousins()
k = kappa(x, leaves(x))
stopifnot(identical(k, c(17/32, 14/32, 1/32)))

# Paternal half brothers with 100\% inbred father
# Genetically indistinguishable from an (outbred) father-son relationship
x = halfSibPed()
founderInbreeding(x, 1) = 1
k = kappa(x, 4:5)
stopifnot(identical(k, c(0, 1, 0)))

}
\seealso{
\code{\link[=condensedIdentity]{condensedIdentity()}}
}
