% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaIbd.R
\name{kappaIbd}
\alias{kappaIbd}
\alias{kappa}
\alias{kappaIbdX}
\title{IBD (kappa) coefficients}
\usage{
kappaIbd(x, ids, sparse = NA, verbose = FALSE)

kappa(x, ids, ...)

kappaIbdX(x, ids, sparse = NA, verbose = FALSE)
}
\arguments{
\item{x}{A pedigree in the form of a \code{\link[pedtools:ped]{pedtools::ped}} object}

\item{ids}{A character (or coercible to character) containing ID labels of
two or more pedigree members.}

\item{sparse}{A positive integer, indicating the pedigree size limit for
using sparse arrays (as implemented by the
\href{https://CRAN.R-project.org/package=slam}{slam} package) instead of
ordinary arrays.}

\item{verbose}{A logical}

\item{...}{Further arguments}
}
\value{
If \code{ids} has length 2: A numeric vector of length 3: \eqn{(\kappa0,
  \kappa1, \kappa2)}

If \code{ids} has length > 2: A data frame with one row for each pair of
individuals, and 5 columns. The first two columns contain the ID labels,
and columns 3-5 contain the IBD coefficients.

For pairs involving inbred individuals (inbred \emph{females} in the X version)
all three coefficients are reported as NA. Furthermore, the X chromosomal
\eqn{kappa2} is NA whenever at least one of the two individuals is male.
}
\description{
Computes the three IBD coefficients summarising the relationship between two
non-inbred individuals. Both autosomal and X chromosomal versions are
implemented. \code{kappa()} is a synonym for \code{kappaIbd()}, but will most likely be
removed (since it conflicts with \code{base::kappa()}).
}
\details{
For non-inbred individuals a and b, their autosomal IBD coefficients
\eqn{(\kappa0, \kappa1, \kappa2)} are defined as follows: \deqn{\kappa_i =
P(a and b share i alleles IBD at a random autosomal locus)}

The autosomal kappa coefficients are computed using the method described by
Karigl (1981) for the condensed identity coefficients. The program first
checks if any of the individuals are inbred; if so, \code{c(NA, NA, NA)} is
returned. If none of the individuals are inbred, Karigl's system of equations
is reduced to only 3 equations: \deqn{\kappa0 + \kappa1 + \kappa2 = 1,}
\deqn{\kappa1 + 2*\kappa2 = 4*\phi_{ab},} \deqn{\kappa1 + 4*\kappa2 =
16*\phi_{ab,ab}.} Here \eqn{\phi_{ab}} is the standard kinship coefficient,
and \eqn{\phi_{ab,ab}} is the generalised kinship coefficient defined by
Karigl (1981). The program calls \code{\link[=generalisedKinship22]{generalisedKinship22()}} to compute this.

The X chromosomal IBD coefficients are defined as in the autosomal case, with
the exception that \eqn{\kappa2} is undefined when at least one of the two
individuals is male. Hence the computation is greatly simplified when males
are involved. Denoting the standard kinship coefficient by \eqn{\phi}, the
formulas are:
\itemize{
\item Both male: \eqn{(\kappa0, \kappa1, \kappa2) = (1-\phi, \phi, NA)}
\item One male, one female: \eqn{(\kappa0, \kappa1, \kappa2) = (1-2*\phi, 2*\phi,
NA)}
\item Two females: As in the autosomal case.
}
}
\examples{
# Siblings
x = nuclearPed(2)
k = kappa(x, 3:4)
stopifnot(identical(k, c(.25, .5, .25)))

# Quad half first cousins
x = quadHalfFirstCousins()
k = kappa(x, leaves(x))
stopifnot(identical(k, c(17/32, 14/32, 1/32)))

# Paternal half brothers with 100\% inbred father
# Genetically indistinguishable from an (outbred) father-son relationship
x = halfSibPed()
founderInbreeding(x, 1) = 1
k = kappa(x, 4:5)
stopifnot(identical(k, c(0, 1, 0)))

}
\references{
G. Karigl (1981). \emph{A recursive algorithm for the calculation of
identity coefficients} Annals of Human Genetics, vol. 45.
}
\seealso{
\code{\link[=kinship]{kinship()}}, \code{\link[=condensedIdentity]{condensedIdentity()}}
}
