% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaVariance.R
\name{kappaVariance}
\alias{kappaVariance}
\alias{deltaVariance}
\title{Variance of realised relatedness coefficients}
\usage{
kappaVariance(x, ids = leaves(x), coef, M = 1)

deltaVariance(x, ids = leaves(x), coef, M = 1)
}
\arguments{
\item{x}{A ped object.}

\item{ids}{A vector naming two members of \code{x}.}

\item{coef}{A single character: one of "k0", "k1", "k2" in \code{kappaVariance},
and one of "D1", "D2", ... "D9" in \code{deltaVariance}.}

\item{M}{A positive number; the chromosome length in Morgan.}
}
\value{
A positive number.
}
\description{
Compute the variance of realised relatedness coefficients, by doubly
integrating the corresponding two-locus coefficients.
}
\details{
The double integral method was used by Guo to compute the variation in genome
sharing (Guo 1995, see also Thompson 2013). The method extends directly to
other coefficients. The implementation here supports Cotterman's kappa
coefficients (of noninbred individuals), and Jacquard's condensed identity
coefficients.

The current implementation is based on stats::integrate and can probably be
optimised in several ways.
}
\examples{
###################################
### Box 1 of Hill & Weir (2011) ###
###################################

# Eq. 4b of Hill & Weir
phi = function(n, l) {
 1/(2*l^2) * (1/4)^n * sum(sapply(1:n, function(r)
   choose(n, r) * (2*r*l - 1 + exp(-2*r*l))/r^2))
}

# Chromosome of 1 Morgan
M = 1

### Full sibs ###

x = nuclearPed(2)
kappaVariance(x, ids = 3:4, coef = "k2", M = M)

# Hill & Weir (Box 1)
16*phi(4,M) - 16*phi(3,M) + 8*phi(2,M) - 2*phi(1,M)


### Double first cousins ###

\dontrun{
dfc = doubleFirstCousins()
kappaVariance(dfc, coef = "k0", M = M)
kappaVariance(dfc, coef = "k1", M = M)
kappaVariance(dfc, coef = "k2", M = M)

# Hill & Weir, Box 1
var_k2 = 64*phi(8,M) - 64*phi(7,M) + 40*phi(6,M) - 20*phi(5,M) +
 33/4*phi(4,M) - 5/2*phi(3,M) + 5/8*phi(2,M)-1/8*phi(1,M)
var_k1 = 4*var_k2
var_k0 = var_k2 + 2 * (4*phi(4,M) - 2*phi(3,M) + 3/4*phi(2,M) - 1/4*phi(1,M))

var_k0
var_k1
var_k2

}

}
\references{
\itemize{
\item Guo (1995) \emph{Proportion of genome shared identical by descent by relatives:
concept, computation, and applications}. Am J Hum Genet.
\item Hill & Weir (2011). \emph{Variation in actual relationship as a consequence of
Mendelian sampling and linkage}. Genet Res.
\item Thompson (2013). \emph{Identity by Descent: Variation in Meiosis, Across
Genomes, and in Populations}. Genetics.
}
}
