% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinpat.R
\name{kinpat}
\alias{kinpat}
\title{Generalised kinship pattern}
\usage{
kinpat(x, pattern, distinct = TRUE)
}
\arguments{
\item{x}{A \code{ped} object}

\item{pattern}{A list of vectors of ID labels.}

\item{distinct}{A logical indicating if different blocks always have distinct
alleles}
}
\value{
An object of class \code{kinpat}.
}
\description{
Generalised kinship pattern
}
\examples{
kinpat(nuclearPed(2), list(1, 3:4))

}
