% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalisedKinship.R
\name{kinPattern}
\alias{kinPattern}
\title{Generalised kinship pattern}
\usage{
kinPattern(x, pattern, internal = F)
}
\arguments{
\item{x}{A \code{ped} object}

\item{pattern}{A list of vectors of ID labels.}

\item{internal}{A logical}
}
\value{
An object of class \code{kinPattern}.
}
\description{
Generalised kinship pattern
}
\examples{
kinPattern(nuclearPed(2), list(1, 3:4))

}
