% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPersonIBD.R
\name{minimalPattern}
\alias{minimalPattern}
\title{Minimal IBD pattern}
\usage{
minimalPattern(x)
}
\arguments{
\item{x}{An integer vector of even length.}
}
\value{
An integer vector of the same length as \code{x}.
}
\description{
Compute the minimal form of given multiperson IBD pattern.
}
\examples{
v = c(1,2,2,3)
stopifnot(identical(minimalPattern(v), c(1,2,1,3)))

}
