% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MULTIPATTERNS_NONINBRED}
\alias{MULTIPATTERNS_NONINBRED}
\title{Multi-person IBD patterns}
\format{A list of 6 matrices. The numbers of rows are respectively 1, 3, 16, 139,
1750 and 29388.}
\usage{
MULTIPATTERNS_NONINBRED
}
\description{
A list of length 6. Each list entry is a matrix of IBD patterns with some
extra information in additional columns. The N'th entry contains all
condensed patterns of N individuals.
}
\keyword{datasets}
