% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribd-package.R
\docType{package}
\name{ribd}
\alias{ribd}
\alias{ribd-package}
\title{ribd: Computation of pedigree-based relatedness coefficients}
\description{
Recursive algorithms for computing various relatedness coefficients,
including Jacquard's condensed identity coefficients. The standard algorithms
are extended to allow inbred founders. Both autosomal and X-linked
coeffiecients are computed.
}
