% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoLocusKinship.R
\name{twoLocusKinship}
\alias{twoLocusKinship}
\title{Two-locus kinship coefficients}
\usage{
twoLocusKinship(x, ids, r, verbose = FALSE)
}
\arguments{
\item{x}{A pedigree in the form of a \code{\link[pedtools:ped]{pedtools::ped}} object}

\item{ids}{A character (or coercible to character) containing ID labels of
two or more pedigree members.}

\item{r}{A number in the interval \eqn{[0, 0.5]}; the recombination rate between the two loci.}

\item{verbose}{A logical}
}
\description{
Computes the two-locus kinship coefficient for two pedigree members, at two
loci separated by a given recombination rate. WORK IN PROGRESS
}
\details{
The implementation is based on the recursive algorithm described by Thompson
(1988).
}
\examples{
# Full sibs
x = nuclearPed(2)

k_0 = twoLocusKinship(x, ids = 3:4, r = 0)
k_0.5 = twoLocusKinship(x, ids = 3:4, r = 0.5)

stopifnot(k_0 == 1/4, k_0.5 == 1/16)

#######################################
# Reproducing Fig. 3 in Thompson (1988)
# Note that in the article, curve (a) is wrong.
# (See published Erratum.)
#######################################

# Pedigrees (a) - (d)
peds = list(
  a = list(ped = linearPed(3), ids = c(1,7)),
  b = list(ped = halfCousinPed(0, 1), ids = c(3,7)),
  c = list(ped = cousinPed(1), ids = c(5,8)),
  d = list(ped = doubleCousins(1, 1, half1 = TRUE, half2 = TRUE),
           ids = c(5,9)))

# Recombination values
rseq = seq(0, 0.5, length = 20)

# Compute two-locus kinship coefficients
kvals = sapply(peds, function(x)
  sapply(rseq, function(r) twoLocusKinship(x$ped, x$ids, r)))

# Plot
matplot(rseq, kvals, type="l")
legend("topright", letters[1:4], col = 1:4, lty = 1:4)

}
\references{
E. A. Thompson (1988). \emph{Two-locus and Three-locus Gene Identity
by Descent in Pedigrees} IMA Journal of Mathematics Applied in Medicine &
Biology, vol. 5.
}
