% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd_identity_x.R
\name{condensedIdentityX}
\alias{condensedIdentityX}
\title{Identity coefficients on X}
\usage{
condensedIdentityX(x, ids, verbose = FALSE, checkAnswer = verbose,
  sparse = Inf)
}
\arguments{
\item{x}{A pedigree object}

\item{ids}{A numeric of length 2 containing ID labels of two pedigree members}

\item{verbose}{A Logical}

\item{checkAnswer}{If TRUE, the result is checked against the output of the
XIBD package, if this is installed. (Ignored if any of the founders are inbred.)}

\item{sparse}{A positive integer, indicating the pedigree size limit for
using sparse arrays (as implemented by the
\href{https://CRAN.R-project.org/package=slam}{slam} package) instead of
ordinary arrays.}
}
\value{
A vector of length 9. Unless both of \code{ids} are female, some entries
are NA.
}
\description{
Computes the X chromosomal condensed identity coefficients of a pairwise
relationship.
}
\details{
The implementation is inspired by Karigl's recursive algorithm (1981) for the
autosomal case, modified to account for X-linked inheritance.
}
\examples{
library(pedtools)

x = fullSibMating(1)
x_sisters = swapSex(x, 5)
x_brothers = swapSex(x, 6)

condensedIdentityX(x, ids = 5:6)
condensedIdentityX(x_sisters, ids = 5:6)
condensedIdentityX(x_brothers, ids = 5:6)

}
