% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd_inbreeding.R
\name{inbreeding}
\alias{inbreeding}
\title{Inbreeding coefficients}
\usage{
inbreeding(x)
}
\arguments{
\item{x}{A pedigree, in the form of a \code{\link[pedtools:ped]{pedtools::ped}} object.}
}
\value{
A vector of length \code{pedsize(x)}.
}
\description{
Compute the inbreeding coefficients of all members of a pedigree. This is a
wrapper of \code{\link[=kinship]{kinship()}} which does the main work. The founders may be
inbred; see \code{\link[pedtools:founderInbreeding]{pedtools::founderInbreeding()}} for how to set this up.
}
\details{
The inbreeding coefficient of a pedigree member is defined as the probability
that, at a random autosomal locus, the two alleles carried by the member are
identical by descent relative to the pedigree. It follows from the definition
that the inbreeding coefficient of a member equals the kinship coefficient of
the parents.
}
\examples{
# Child of half siblings: f = 1/8
x = halfCousinPed(0, child = TRUE)
inbreeding(x)

# If the father is 100\% inbred, the inbreeding coeff of the child doubles
founderInbreeding(x, 1) = 1
inbreeding(x)

}
\seealso{
\code{\link[=kinship]{kinship()}}
}
