% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffTable.R
\name{kin2deg}
\alias{kin2deg}
\title{Degree of relationship}
\usage{
kin2deg(kin, unrelated = Inf)
}
\arguments{
\item{kin}{A vector of kinship coefficients, i.e., numbers in \verb{[0, 1]}.}

\item{unrelated}{The conversion of unrelated individuals (\code{kin = 0}).
Mathematically this corresponds to \code{degree = Inf}, but in some situations
\code{degree = NA} or something else might be preferable.}
}
\value{
An integer vector of the same length as \code{kin}.
}
\description{
Converts a vector of kinship coefficients to "degrees of relationship", as
used by some software for relatedness inference (e.g. KING).
}
\details{
The implementation uses the conversion formula \deqn{deg = round(-log2(kin) -
1).}
}
\examples{
x = cousinPed(1)

# Kinship matrix
k = kinship(x)

# Degrees
deg = kin2deg(k)
deg

# First cousins are 3rd degree
stopifnot(deg['7', '8'] == 3)

}
\seealso{
\code{\link[=kinship]{kinship()}}, \code{\link[=coeffTable]{coeffTable()}}
}
