% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribd-package.R
\docType{package}
\name{ribd-package}
\alias{ribd}
\alias{ribd-package}
\title{ribd: Pedigree-based Relatedness Coefficients}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Recursive algorithms for computing various relatedness coefficients, including pairwise kinship, kappa and identity coefficients. Both autosomal and X-linked coefficients are computed. Founders are allowed to be inbred, which enables construction of any given kappa coefficients, as described in Vigeland (2020) \doi{10.1007/s00285-020-01505-x}. In addition to the standard coefficients, 'ribd' also computes a range of lesser-known coefficients, including generalised kinship coefficients (Karigl (1981) \doi{10.1111/j.1469-1809.1981.tb00341.x}; Weeks and Lange (1988) \url{https://pubmed.ncbi.nlm.nih.gov/3422543/}), two-locus coefficients (Thompson (1988) \doi{10.1093/imammb/5.4.261}) and multi-person coefficients. Many features of 'ribd' are available through the online app 'QuickPed' at \url{https://magnusdv.shinyapps.io/quickped}; see Vigeland (2022) \doi{10.1186/s12859-022-04759-y}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/ribd}
  \item \url{https://magnusdv.github.io/pedsuite/}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
