% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoLocusIBD.R
\name{twoLocusIBD}
\alias{twoLocusIBD}
\title{Two-locus IBD coefficients}
\usage{
twoLocusIBD(x, ids, r)
}
\arguments{
\item{x}{A pedigree in the form of a \code{\link[pedtools:ped]{pedtools::ped}} object.}

\item{ids}{A character (or coercible to character) containing ID labels of
two pedigree members.}

\item{r}{A number in the interval \eqn{[0, 0.5]}; the recombination rate
between the two loci.}
}
\description{
Computes the two-locus IBD coefficient \eqn{\kappa_{1,1}}{k_11} of a pair of
pedigree members, at a given recombination rate.
}
\details{
Let A, B be two pedigree members, and L1, L2 two loci with a given
recombination rate r. The two-locus IBD coefficient
\eqn{\kappa_{1,1}(r)}{k_11(r)} is defined as the probability that A and B
have IBD status 1 at both L1 and L2 simultaneously. Note that it is not
required that the IBD alleles are in cis (or in trans for that matter), in
either individual.

The method of computation is via the two-locus kinship coefficients involving
the parents of A and B.
}
\examples{
######################################
# A classic example of three relationships with the same one-locus IBD
# coefficients, but different two-locus IBD coefficients. As a consequence,
# these relationships cannot be separated using unlinked markers, but they
# are separable (in theory) using pairs of linked markers.
######################################
peds = list(
  GrandP  = list(ped = linearPed(2),    ids = c(1,5)),
  Uncle   = list(ped = cousinPed(0, 1), ids = c(3,6)),
  HalfSib = list(ped = halfSibPed(),    ids = c(4,5)))

# Recombination values
rseq = seq(0, 0.5, length = 20)

# Compute two-locus IBD coefficients (k11)
# kvals = sapply(peds, function(x)
#    sapply(rseq, function(r) twoLocusIBD(x$ped, x$ids, r)))

# Plot
# matplot(rseq, kvals, type="l")
# legend("topright", names(peds), col = 1:3, lty = 1:3)

}
\seealso{
\code{\link[=twoLocusKinship]{twoLocusKinship()}}
}
